patterns/strands-single-agent/requirements.txt
strands-agents
boto3
bedrock-agentcore[strands-agents]
mcp
requests
-e ../../.
